Fixed some static code analysis warnings

+++ include/gracht/types.h
@@ -27,8 +27,8 @@
 #include <stddef.h>
 
 #if defined(_WIN32)
-typedef void*        gracht_handle_t;
-typedef unsigned int gracht_conn_t;
+typedef void*      gracht_handle_t;
+typedef uintptr_t  gracht_conn_t;
 #define GRACHT_HANDLE_INVALID NULL
 #define GRACHT_CONN_INVALID   (unsigned int)0
 #elif defined(MOLLENOS)
